Initial Networking
I need to add a local mode to skip over the networking portion.

To Host a game

Host = null (Just press okay)
Player id= anything
Press yes when it asks you to start a server.

To join a game
Host = IP of Server
Player ID = If it's the same as the server you're going to to have
problems.