sheeeeIIITTTT
characters now save.
Also fixed the HudObject memory leak in CharSetupPanel (whoopsy).

added a method to MPanel which prints each unique action type in the
hudOb array, and prints the number of instances of each action type.
This is useful in finding any memory leaks related to the repainting of
HudObjects (if the number of each action is incrementing inexplicably)