fixed clicking
the method call myClick.getClicked(); not only gets a boolean as to
whether or not the mouse was clicked, but then clears the flag that it
uses to determine if the mouse was clicked. So, if you require two
checks of getClicked in a row, it would NEVER pass.

This wasn't an issue until there was a super.checkClick() call, where
both the child and the super would call that method, the super would
reset the flag, and the child would stumble on it every time. This is
why having the click check in the MPanel's checkClick method broke the
clicking in leveldesigner and gamepanel, but not in networksetup panel.

from now on, all children of MPanel should use MPanel's "anyClicked"
boolean to see if it has an unhandled click.