Implemented Seeds and upgraded Main Menu
Levels now specify the starting location of players with tile "seeds"
which are assigned in the xml file.

Main Menu is now more organized.

Level designer does not have a way of assigning seeds or curtains, so
that will need to be upgraded at a later date.

Character creator still does not confirm to new stat system, also needs
changed.