kaBOOM! SKIRMISH PANEL
Fixed highlight issue. Also added MIntegers, which are onscreenobject
representation of any positive integer. This pulls from a number
spreadsheet (45.png in pics/huds/numbers) This is very rough atm because
the spritesheet picture isn't made of consistently sized parts, so it
currently has a cookie-cutter visual effect. This will go away when we
have a real, good, spreadsheet.