Network Messages
Added in the base code for the network message. I also added in a
Base64Encoder (serialization). I don't know if we are going to need that
package for anything else, but for know it's part of the Network
package.

Next up is to create the controllers to handle these messages as well as
the updates to the server code to handle the lobby creation.